requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
Pillow>=10.0.0
PyYAML>=6.0.1
streamlit>=1.33.0
opencv-python-headless>=4.9.0
scikit-image>=0.22.0 